Refactor tx-status-request to addon with switch in public_settings.h
