+++ src/qubic.cpp
@@ -36,8 +36,7 @@
 
 #include "tick_storage.h"
 
-#include "tx_status_request.h"
-
+#include "addons/tx_status_request.h"
 
 ////////// Qubic \\\\\\\\\\
 
@@ -1431,12 +1430,14 @@ static void requestProcessor(void* ProcedureArgument)
                 }
                 break;
 
+#if ADDON_TX_STATUS_REQUEST
                 /* qli: process RequestTxStatus message */
                 case REQUEST_TX_STATUS:
                 {
                     processRequestConfirmedTx(processorNumber, peer, header);
                 }
                 break;
+#endif
 
                 }
 
@@ -2098,7 +2099,9 @@ static void processTick(unsigned long long processorNumber)
     if (nextTickData.epoch == system.epoch)
     {
         auto* tsCurrentTickTransactionOffsets = ts.tickTransactionOffsets.getByTickIndex(tickIndex);
+#if ADDON_TX_STATUS_REQUEST
         tickTxIndexStart[system.tick - system.initialTick] = numberOfTransactions; // qli: part of tx_status_request add-on
+#endif
         bs->SetMem(entityPendingTransactionIndices, sizeof(entityPendingTransactionIndices), 0);
         // reset solution task queue
         score->resetTaskQueue();
@@ -2166,18 +2169,26 @@ static void processTick(unsigned long long processorNumber)
                         entityPendingTransactionIndices[spectrumIndex] = 1;
 
                         numberOfTransactions++;
+#if ADDON_TX_STATUS_REQUEST
                         tickTxIndexStart[system.tick - system.initialTick + 1] = numberOfTransactions; // qli: part of tx_status_request add-on
+#endif
                         if (decreaseEnergy(spectrumIndex, transaction->amount))
                         {
                             increaseEnergy(transaction->destinationPublicKey, transaction->amount);
                             
                             if (transaction->amount)
                             {
+#if ADDON_TX_STATUS_REQUEST
                                 saveConfirmedTx(numberOfTransactions - 1, 1, system.tick, nextTickData.transactionDigests[transactionIndex]); // qli: save tx
+#endif
                                 const QuTransfer quTransfer = { transaction->sourcePublicKey , transaction->destinationPublicKey , transaction->amount };
                                 logQuTransfer(quTransfer);
-                            }else{
+                            }
+                            else
+                            {
+#if ADDON_TX_STATUS_REQUEST
                                 saveConfirmedTx(numberOfTransactions - 1, 0, system.tick, nextTickData.transactionDigests[transactionIndex]); // qli: save tx
+#endif
                             }
 
                             if (isZero(transaction->destinationPublicKey))
@@ -2488,7 +2499,9 @@ static void processTick(unsigned long long processorNumber)
                         }
                         else
                         {
+#if ADDON_TX_STATUS_REQUEST
                             saveConfirmedTx(numberOfTransactions - 1, 0, system.tick, nextTickData.transactionDigests[transactionIndex]); // qli: save tx
+#endif
                         }
                     }
                 }
@@ -2715,7 +2728,9 @@ static void beginEpoch1of2()
 #ifndef NDEBUG
     ts.checkStateConsistencyWithAssert();
 #endif
+#if ADDON_TX_STATUS_REQUEST
     beginEpochTxStatusRequestAddOn(system.initialTick);
+#endif
 
     for (unsigned int i = 0; i < SPECTRUM_CAPACITY; i++)
     {
@@ -4007,12 +4022,13 @@ static bool initialize()
         if (!initLogging())
             return false;
 
-        // qli: init tx_status_request
+#if ADDON_TX_STATUS_REQUEST
         if (!initTxStatusRequestAddOn())
         {
             logToConsole(L"initTxStatusRequestAddOn() failed!");
             return false;
         }
+#endif
 
         logToConsole(L"Loading system file ...");
         bs->SetMem(&system, sizeof(system), 0);
@@ -4254,7 +4270,9 @@ static void deinitialize()
 
     deinitLogging();
 
+#if ADDON_TX_STATUS_REQUEST
     deinitTxStatusRequestAddOn();
+#endif
 
     for (unsigned int processorIndex = 0; processorIndex < MAX_NUMBER_OF_PROCESSORS; processorIndex++)
     {
